Start #1: Middleware base class setup, Transport inherits from Middleware. No real work has been done to integrate the middleware concept into Agency, other than adding a `middlewares` attribute. Ideally, middlewares will be defined through a class configurator, such as `Agency.configure{ |config| config.use_middleware SomeMiddlewareClass }`.
